ipfs context fully implemented and working
