+++ my-app/src/contexts/IPFScontext.js
@@ -11,6 +11,30 @@ export const useIPFScontext = () =>{
 
 
 export const IPFSprovider = ({children}) =>{
+    async function addToIpfs(content) {
+        try {
+    
+          const addedData = await ipfs.add({ content });
+    
+          return addedData;
+        } catch (error) {
+          console.error("An error occurred while adding to IPFS:", error);
+          throw error; 
+        }
+      }
+
+    async function fetchFromIpfs(ipfsHash) {
+        let stringData = '';
+        for await (const chunk of ipfs.cat(ipfsHash)) {
+            stringData += new TextDecoder().decode(chunk);
+        }
+        try {
+            return JSON.parse(stringData);
+        } catch (error) {
+            console.error("Error parsing JSON from IPFS:", error, "stringData:", stringData);
+            throw error;
+        }
+    }
     
     const auth = 'Basic ' + Buffer.from(process.env.NEXT_PUBLIC_INFURA_PROJECTID + ':' + process.env.NEXT_PUBLIC_INFURA_IPFS).toString('base64');
 
@@ -29,6 +53,8 @@ export const IPFSprovider = ({children}) =>{
         <IPFScontext.Provider
         value ={{
             ipfs,
+            fetchFromIpfs,
+            addToIpfs,
         }}
         >
         {children}
