+++ my-app/src/pages/_app.js
@@ -8,6 +8,7 @@ import { Web3Provider } from "@/contexts/Web3Context";
 import { DataBaseProvider } from "@/contexts/DataBaseContext";
 import { LeaderboardProvider } from "@/contexts/leaderboardContext";
 import { VotingProvider } from "@/contexts/votingContext";
+import { IPFSprovider } from "@/contexts/IPFScontext";
 
 const theme = extendTheme({
   config: {
@@ -24,28 +25,30 @@ function MyApp({ Component, pageProps }) {
     document.title = 'KUBI DAO';
   }, []);
   return (
-    <Web3Provider>
-      <DataBaseProvider>
-        <VotingProvider>
-        <LeaderboardProvider>
-          <ChakraProvider theme={theme}>
-            <Flex 
-              direction="column" 
-              minH="100vh" 
-              bg={router.pathname === "/tasks" ? "clear" : "#1285ff"} // Set the background based on the current route
-            >
-              <Box as="header">
-                <NavBar />
-              </Box>
-              <Flex as="main" direction="column" flex="1" overflow="hidden">
-                <Component {...pageProps} />
+    <IPFSprovider>
+      <Web3Provider>
+        <DataBaseProvider>
+          <VotingProvider>
+          <LeaderboardProvider>
+            <ChakraProvider theme={theme}>
+              <Flex 
+                direction="column" 
+                minH="100vh" 
+                bg={router.pathname === "/tasks" ? "clear" : "#1285ff"} // Set the background based on the current route
+              >
+                <Box as="header">
+                  <NavBar />
+                </Box>
+                <Flex as="main" direction="column" flex="1" overflow="hidden">
+                  <Component {...pageProps} />
+                </Flex>
               </Flex>
-            </Flex>
-          </ChakraProvider>
-        </LeaderboardProvider>
-        </VotingProvider>
-      </DataBaseProvider>
-    </Web3Provider>
+            </ChakraProvider>
+          </LeaderboardProvider>
+          </VotingProvider>
+        </DataBaseProvider>
+      </Web3Provider>
+    </IPFSprovider>
   );
 }
